backend/pyproject.toml

[tool.poetry]
name = "fastapi-chatbot-backend"
version = "0.1.0"
description = "FastAPI backend for the TypeScript chatbot application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
openai = "^1.3.7"
python-dotenv = "^1.0.0"
pydantic = "^2.9.2"
python-multipart = "^0.0.6"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn main_sqlite:app --reload"
